@@ -12,17 +12,17 @@
   "private": true,
   "dependencies": {
     "@angular-eslint/eslint-plugin": "^17.1.1",
-    "@angular/animations": "^17.1.1",
+    "@angular/animations": "^18.1.0",
     "@angular/cdk": "^17.3.2",
-    "@angular/common": "^17.1.1",
-    "@angular/compiler": "^17.1.1",
-    "@angular/core": "^17.1.1",
+    "@angular/common": "^18.1.0",
+    "@angular/compiler": "^18.1.0",
+    "@angular/core": "^18.1.0",
     "@angular/fire": "^17.0.1",
-    "@angular/forms": "^17.1.1",
+    "@angular/forms": "^18.1.0",
     "@angular/material": "^17.3.2",
-    "@angular/platform-browser": "^17.1.1",
-    "@angular/platform-browser-dynamic": "^17.1.1",
-    "@angular/router": "^17.1.1",
+    "@angular/platform-browser": "^18.1.0",
+    "@angular/platform-browser-dynamic": "^18.1.0",
+    "@angular/router": "^18.1.0",
     "@fontsource/open-sans": "^5.0.22",
     "@fontsource/playpen-sans": "^5.0.2",
     "dayjs": "^1.11.10",
@@ -36,9 +36,9 @@
     "zone.js": "~0.14.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.1.1",
-    "@angular/cli": "~17.1.1",
-    "@angular/compiler-cli": "^17.1.1",
+    "@angular-devkit/build-angular": "^18.1.0",
+    "@angular/cli": "~18.1.0",
+    "@angular/compiler-cli": "^18.1.0",
     "@types/feather-icons": "^4.29.2",
     "@types/jasmine": "~4.3.0",
     "@types/ramda": "^0.29.9",
@@ -55,6 +55,6 @@
     "postcss": "^8.4.29",
     "tailwind-merge": "^2.2.2",
     "tailwindcss": "^3.3.3",
-    "typescript": "~5.2.2"
+    "typescript": "~5.5.3"
   }
 }
